additional light sources

+++ src/render/node/chunk-node/chunk-node.ts
@@ -41,7 +41,7 @@ export class ChunkNode {
 		const output = new Texture();
 		const normal = new Texture(undefined, undefined, null, gl.RGBA16F, gl.RGBA, gl.FLOAT);
 		const position = new Texture(undefined, undefined, null, gl.RGBA32F, gl.RGBA, gl.FLOAT);
-		this.chunks = new Texture(4096, 1, undefined, gl.RGBA32F, gl.RGBA, gl.FLOAT);
+		this.chunks = new Texture(4096, 1, undefined, gl.RGBA16F, gl.RGBA, gl.FLOAT);
 		this.colors = new Texture(4096, 1);
 
 		this.frameBuffer = new FrameBuffer([output, normal, position], true, true);
